backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
aiosqlite==0.20.0
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.4.1
xgboost==2.0.3
joblib==1.3.2
pandas==2.2.0
numpy==1.26.4
tldextract==5.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-dotenv==1.0.1
